move plotting functions into module (#68)
